Move recs handling to systemManager + better names
