+++ client/src/three/systems/types.ts
@@ -1,5 +1,6 @@
 import { HexPosition } from "@/types";
 import { ID, StructureType } from "@bibliothecadao/eternum";
+import { StructureProgress } from "../scenes/constants";
 
 export type ArmySystemUpdate = {
   entityId: ID;
@@ -12,6 +13,7 @@ export type StructureSystemUpdate = {
   hexCoords: HexPosition;
   structureType: StructureType;
   isMine: boolean;
+  stage: StructureProgress;
 };
 
 export type TileSystemUpdate = {
